Introduce WeightExporter and move weights send group methods
