+++ python/sglang/srt/model_executor/model_runner.py
@@ -153,6 +153,7 @@ from sglang.srt.model_executor.piecewise_cuda_graph_runner import (
     PiecewiseCudaGraphRunner,
 )
 from sglang.srt.model_executor.pool_configurator import MemoryPoolConfig
+from sglang.srt.model_executor.weight_exporter import WeightExporter
 from sglang.srt.model_executor.weight_updater import WeightUpdater
 from sglang.srt.model_loader.loader import get_model_loader
 from sglang.srt.model_loader.remote_instance_weight_loader_utils import (
@@ -187,7 +188,6 @@ from sglang.srt.utils import (
     get_available_gpu_memory,
     get_bool_env_var,
     init_cublas,
-    init_custom_process_group,
     is_hip,
     is_host_cpu_arm64,
     is_npu,
@@ -555,7 +555,12 @@ class ModelRunner(ModelRunnerKVCacheMixin):
 
         # For weight updates
         self.weight_updater = WeightUpdater(tp_rank=self.tp_rank, model_runner_ref=self)
-        self._weights_send_group = {}
+        self.weight_exporter = WeightExporter(
+            tp_rank=self.tp_rank,
+            tp_size=self.tp_size,
+            gpu_id=self.gpu_id,
+            model_runner_ref=self,
+        )
 
     def _build_model_config(
         self, server_args, model_path=None, model_revision=None, is_draft_model=False
@@ -1685,103 +1690,6 @@ class ModelRunner(ModelRunnerKVCacheMixin):
             "No healthy rank found for broadcasting expert location metadata. "
             "All ranks are marked as elastic_ep_rejoin."
         )
-
-    def init_weights_send_group_for_remote_instance(
-        self,
-        master_address,
-        ports,
-        group_rank,
-        world_size,
-        group_name,
-        backend="nccl",
-    ):
-        assert (
-            torch.distributed.is_initialized()
-        ), "Default torch process group must be initialized"
-        assert group_name != "", "Group name cannot be empty"
-
-        ports_list = ports.split(",")
-        assert (
-            len(ports_list) == self.tp_size
-        ), f"Expected {self.tp_size} ports, but got {len(ports_list)} ports."
-        group_port = ports_list[self.tp_rank]
-        group_name = f"{group_name}_{group_port}_{self.tp_rank}"
-
-        logger.info(
-            f"init custom process group: tp_rank={self.tp_rank}, gpu_id={self.gpu_id}, master_address={master_address}, master_port={group_port}, "
-            f"group_rank={group_rank}, world_size={world_size}, group_name={group_name}, backend={backend}"
-        )
-
-        torch.cuda.empty_cache()
-        success = False
-        message = ""
-        try:
-            na = NetworkAddress(master_address, group_port)
-            self._weights_send_group[group_name] = init_custom_process_group(
-                backend=backend,
-                init_method=na.to_tcp(),
-                world_size=world_size,
-                rank=group_rank,
-                group_name=group_name,
-                device_id=torch.device("cuda", self.gpu_id),
-            )
-            dist.barrier(group=self._weights_send_group[group_name])
-            success = True
-            message = f"Succeeded to init group through {na.to_host_port_str()} group."
-        except Exception as e:
-            message = f"Failed to init group: {e}."
-            logger.error(message)
-
-        torch.cuda.empty_cache()
-        return success, message
-
-    def send_weights_to_remote_instance(
-        self,
-        master_address,
-        ports,
-        group_name,
-    ):
-        assert (
-            torch.distributed.is_initialized()
-        ), "Default torch process group must be initialized"
-        assert group_name != "", "Group name cannot be empty"
-
-        ports_list = ports.split(",")
-        assert (
-            len(ports_list) == self.tp_size
-        ), f"Expected {self.tp_size} ports, but got {len(ports_list)} ports."
-        group_port = ports_list[self.tp_rank]
-        group_name = f"{group_name}_{group_port}_{self.tp_rank}"
-
-        if self._weights_send_group[group_name] is not None:
-            send_group = self._weights_send_group[group_name]
-        else:
-            message = f"Group {group_name} not in _weights_send_group list. Please call `init_weights_send_group_for_remote_instance` first."
-            logger.error(message)
-            return False, message
-
-        torch.cuda.empty_cache()
-        success = False
-        na = NetworkAddress(master_address, group_port)
-        message = ""
-        try:
-            for _, weights in self.model.named_parameters():
-                torch.distributed.broadcast(
-                    weights,
-                    src=0,
-                    group=send_group,
-                )
-            success = True
-            message = f"Succeeded to send weights through {na.to_host_port_str()} {group_name}."
-        except Exception as e:
-            message = f"Failed to send weights: {e}."
-            logger.error(message)
-
-        # destroy the process group after sending weights
-        del self._weights_send_group[group_name]
-        torch.distributed.distributed_c10d.destroy_process_group(send_group)
-        torch.cuda.empty_cache()
-        return success, message
 
     def get_weights_by_name(
         self, name: str, truncate_size: int = 100
